requirements.txt
pandas
plotly
openpyxl
xlsxwriter
fpdf2
google-generativeai
